refactor: separate the personal hygiene items types from the types file
